requirements.txt
PyGithub 1.55
discord.py 1.7.3